Move sys stderr test to its own space since it doesn't test server
